Cricket-API-main/Cricket-API-main/requirements.txt
Flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
googlesearch-python==1.3.0
flask-cors==4.0.0
python-dotenv==1.0.0
lxml==4.9.3